use int for amoebalifespan

+++ Menu/Components/OnlineSlugcatAbilitiesInterface.cs
@@ -297,7 +297,7 @@ namespace RainMeadow.UI.Components
 
 
                 //Amoeba duration
-                amoebaLifespanTextBox = new(new Configurable<float>(RainMeadow.rainMeadowOptions.AmoebaDuration.Value), positioner - spacing * 4 + new Vector2(-7.5f, 0), 40)
+                amoebaLifespanTextBox = new(new Configurable<int>(RainMeadow.rainMeadowOptions.AmoebaDuration.Value), positioner - spacing * 4 + new Vector2(-7.5f, 0), 40)
                 {
                     alignment = FLabelAlignment.Center,
                     description = menu.Translate("Amoeba duration time in seconds")
